pre-v0.9.3
MISRA-C:2004 Rules are validated.

+++ boos/system/boos.system.Main.c
@@ -51,7 +51,7 @@ static int8 systemInit(void)
  * 
  * @return error code or zero.
  */   
-int16 main(void)
+int main(void)
 {
   return systemInit();
 }
